fixing up errors and creating the dbcontext for sql database

+++ ProgPart1/Models/HR.cs
@@ -8,7 +8,7 @@ namespace ProgPart1.Models
         {
             public string FirstName { get; set; }
             public string LastName { get; set; }
-            public int HourlyRate { get; set; }
+            public decimal HourlyRate { get; set; }
         }
 
     }
